mtd-utils: make hashtable_iterator functions static inline
GCC 5 changed the default C standard to gnu11 which changed the behavior of the
inline keyword. This causes mtd-utils to fail to build with duplicate
definition errors involving hashtable_iterator_key and
hashtable_iterator_value.

This patch fixes the build by making the functions static inline and removing
the extra definitions in hashtable_itr.c.

Signed-off-by: James Cowgill <james410@cowgill.org.uk>